+++ packages/terra-button-group/src/ButtonGroup.jsx
@@ -170,7 +170,10 @@ class ButtonGroup extends React.Component {
 
     React.Children.forEach(children, (child, index) => {
       const isSelected = selectedKeys.indexOf(child.key) > -1;
-      const tabAttr = (this.props.onChange && !isMultiSelect) ? { tabIndex: (isSelected) ? '0' : '-1' } : null;
+      // Checking whether any option is selected or not
+      const isAnyOptionSelected = children && Array.isArray(children) ? children.findIndex((childItem) => this.props.selectedKeys.includes(childItem?.key)) : -1;
+      const shouldTabIndexBeZero = (onChange && btnRole === 'radio' && isSelected) || (btnRole === 'radio' && isAnyOptionSelected < 0 && index === 0);
+      const tabAttr = (onChange && !isMultiSelect) && { tabIndex: shouldTabIndexBeZero ? '0' : '-1' };
       const cloneChild = React.cloneElement(child, {
         role: btnRole || undefined,
         ...tabAttr,