+++ packages/terra-core-docs/src/terra-dev-site/doc/progress-bar/example/ProgressBarColor.jsx
@@ -7,13 +7,13 @@ const cx = classNames.bind(styles);
 
 const ProgressBarColor = () => (
   <div>
-    <ProgressBar value={5} max={10} colorClass={cx(['color-bar-example-1'])} />
+    <ProgressBar valueText="5%" value={5} max={10} colorClass={cx(['color-bar-example-1'])} />
     <br />
     <br />
-    <ProgressBar value={7.5} max={10} colorClass={cx(['color-bar-example-2'])} />
+    <ProgressBar valueText="7.5%" value={7.5} max={10} colorClass={cx(['color-bar-example-2'])} />
     <br />
     <br />
-    <ProgressBar value={10} max={10} colorClass={cx(['color-bar-example-3'])} />
+    <ProgressBar valueText="10%" value={10} max={10} colorClass={cx(['color-bar-example-3'])} />
   </div>
 );