+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/Examples.2/SearchFieldFocus.jsx
@@ -15,6 +15,8 @@ const SearchFieldFocus = () => {
     <div>
       <Button text="Focus Me" onClick={handleButtonClick} id="search-field-focus-button" />
       <SearchField
+        isLabelVisible
+        groupName="Find a clinic"
         inputRefCallback={(inputRef) => { searchElement.current = inputRef; }}
       />
     </div>