+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/Examples.2/SearchFieldProgrammaticSet.jsx
@@ -56,6 +56,8 @@ const SearchFieldProgrammaticSet = () => {
         minimumSearchTextLength={3}
         value={text}
         inputRefCallback={(ref) => { inputRef.current = ref; }}
+        isLabelVisible
+        groupName="Find a clinic"
       />
       <div id="search-callback-text">
         {message}