+++ packages/terra-core-docs/src/terra-dev-site/test/search-field/SearchFieldDefaultValue.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import SearchField from 'terra-search-field';
 
-export default () => <SearchField id="searchfield" defaultValue="Default" />;
+export default () => <SearchField id="searchfield" defaultValue="Default" isLabelVisible groupName="Find a clinic" />;