+++ packages/terra-dropdown-button/tests/jest/DropdownList.test.jsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 import { IntlProvider } from 'react-intl';
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
 import translationsFile from '../../translations/en.json';
-/* eslint-disable-next-line import/no-extraneous-dependencies */
 import DropdownList from '../../src/_DropdownList';
 import { Item } from '../../src/DropdownButton';
 import Util from '../../src/_DropdownListUtil';
@@ -31,7 +28,7 @@ afterEach(() => {
 
 describe('Dropdown List', () => {
   it('renders a default dropdown list', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
       </DropdownList>,
@@ -44,7 +41,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list with a set width', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <DropdownList requestClose={() => {}} width="440px">
         <Item label="1st Option" onSelect={() => {}} />
       </DropdownList>,
@@ -53,7 +50,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list with multiple children', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -64,7 +61,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list a non-default focused option', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -77,7 +74,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list an active option', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -90,7 +87,7 @@ describe('Dropdown List', () => {
   });
 
   it('renders a dropdown list an active and focused option', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <DropdownList requestClose={() => {}}>
         <Item label="1st Option" onSelect={() => {}} />
         <Item label="2nd Option" onSelect={() => {}} />
@@ -103,7 +100,7 @@ describe('Dropdown List', () => {
   });
 
   it('correctly applies the theme context className', () => {
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
         <DropdownList requestClose={() => {}}>
           <Item label="1st Option" onSelect={() => {}} />
@@ -119,7 +116,7 @@ describe('Dropdown List', () => {
     //  Sets the mock return value for isMac
     jest.spyOn(Util, 'isMac').mockReturnValue(true);
 
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <IntlProvider locale="en" messages={translationsFile}>
         <DropdownList id="dropdownList" requestClose={() => { }}>
           <Item id="firstItem" label="1st Option" onSelect={() => { }} />
@@ -145,7 +142,7 @@ describe('Dropdown List', () => {
     //  Sets the mock return value for isMac
     jest.spyOn(Util, 'isMac').mockReturnValue(false);
 
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <IntlProvider locale="en" messages={translationsFile}>
         <DropdownList id="dropdownList" requestClose={() => { }}>
           <Item id="firstItem" label="1st Option" onSelect={() => { }} />