+++ packages/terra-form-select/tests/jest/SearchSelectField.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
 import SearchSelectField from '../../src/SearchSelectField';
 
 it('should render SearchSelectField component with max height', () => {
@@ -14,6 +12,6 @@ it('should render SearchSelectField component with max height', () => {
     </SearchSelectField>
   );
 
-  expect(shallowWithIntl(select).dive()).toMatchSnapshot();
+  expect(enzymeIntl.shallowWithIntl(select).dive()).toMatchSnapshot();
 });