+++ packages/terra-scroll/tests/jest/Scroll.test.jsx
@@ -3,11 +3,11 @@ import Scroll from '../../lib/Scroll';
 
 // Snapshot Tests
 it('should render a default component', () => {
-  const scrollable = shallow(<Scroll />);
+  const scrollable = enzyme.shallow(<Scroll />);
   expect(scrollable).toMatchSnapshot();
 });
 
 it('should render with props', () => {
-  const scrollable = shallow(<Scroll refCallback={jest.fn()}>child content</Scroll>);
+  const scrollable = enzyme.shallow(<Scroll refCallback={jest.fn()}>child content</Scroll>);
   expect(scrollable).toMatchSnapshot();
 });