+++ packages/terra-status/tests/jest/Status.test.jsx
@@ -14,39 +14,39 @@ const arrange = <Arrange fitStart={image} fill={simpleText} />;
 // Snapshot Tests
 it('should render a image with status', () => {
   const imageWithStatus = <Status>{image}</Status>;
-  const wrapper = shallow(imageWithStatus);
+  const wrapper = enzyme.shallow(imageWithStatus);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render a icon with status', () => {
   const iconWithStatus = <Status>{icon}</Status>;
-  const wrapper = shallow(iconWithStatus);
+  const wrapper = enzyme.shallow(iconWithStatus);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render a simpleText with status', () => {
   const simpleTextWithStatus = <Status>{simpleText}</Status>;
-  const wrapper = shallow(simpleTextWithStatus);
+  const wrapper = enzyme.shallow(simpleTextWithStatus);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render a simpleText with correct class', () => {
   const simpleTextWithStatus = <Status colorClass="testClass">{simpleText}</Status>;
-  const wrapper = shallow(simpleTextWithStatus);
+  const wrapper = enzyme.shallow(simpleTextWithStatus);
   expect(wrapper).toMatchSnapshot();
   expect(wrapper.getElement(0).props.className).toContain('testClass');
 });
 
 it('should render a arrange with status and customProps', () => {
   const arrangeWithStatus = <Status id="id">{arrange}</Status>;
-  const wrapper = shallow(arrangeWithStatus);
+  const wrapper = enzyme.shallow(arrangeWithStatus);
   expect(wrapper).toMatchSnapshot();
 });
 
 // Structure test
 it('should have indicator section of component with correct class, style and type', () => {
   const arrangeWithStatus = <Status className="testClass">{arrange}</Status>;
-  const wrapper = shallow(arrangeWithStatus);
+  const wrapper = enzyme.shallow(arrangeWithStatus);
   expect(wrapper.getElement(0).props.className).toContain('status');
   expect(wrapper.getElement(0).props.className).toContain('testClass');
 });
@@ -54,6 +54,6 @@ it('should have indicator section of component with correct class, style and typ
 // No Color Provided
 it('should render status with no style', () => {
   const statusNoColor = <Status>{simpleText}</Status>;
-  const wrapper = render(statusNoColor);
+  const wrapper = enzyme.render(statusNoColor);
   expect(wrapper).toMatchSnapshot();
 });