+++ packages/terra-tag/tests/jest/TagList.test.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
 import TagList from '../../src/TagList';
 
 describe('TagList Component', () => {
   it('should render tagList with 2 items"', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <TagList ariaLabel="Tag List">
         <div id="tag1">Tag 1</div>
         <div id="tag2">Tag 2</div>
@@ -15,7 +13,7 @@ describe('TagList Component', () => {
   });
 
   it('should render correctly with custom props', () => {
-    const wrapper = shallowWithIntl(
+    const wrapper = enzymeIntl.shallowWithIntl(
       <TagList
         ariaLabel="Custom Tag List"
         ariaLabelledBy="label-id"