+++ packages/terra-avatar/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 3.20.0 - (February 20, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 3.19.0 - (February 15, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 3.18.0 - (November 13, 2023)
 
 * Changed