+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 3.66.0 - (February 20, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 3.65.0 - (February 15, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 3.64.0 - (November 13, 2023)
 
 * Changed