+++ packages/terra-form-select/CHANGELOG.md
@@ -6,6 +6,11 @@
   * Changed --terra-form-select-option-padding value to match with the orion-fusion-theme for consistency.
   * Changed onHover style.
 
+## 6.55.0 - (February 15, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 6.54.0 - (January 9, 2024)
 
 * Changed