+++ packages/terra-section-header/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 2.67.0 - (February 23, 2024)
+
+* Changed
+  * Allow the title of the `terra-section-header` to be sticky. 
+
+## 2.66.0 - (February 15, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 2.65.1 - (December 11, 2023)
 
 * Fixed