+++ packages/terra-status-view/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 4.73.0 - (February 20, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 4.72.0 - (February 15, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 4.71.0 - (January 9, 2024)
 
 * Changed